requirements/base.txt
# ─── Core ML & NLP ────────────────────────────────────────────────────────────
sentence-transformers>=2.6.1
transformers>=4.40.0
torch>=2.2.0
tokenizers>=0.19.0

# ─── Vector Search ────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0
pgvector>=0.2.5

# ─── Clustering ───────────────────────────────────────────────────────────────
hdbscan>=0.8.38
scikit-learn>=1.4.2
umap-learn>=0.5.6

# ─── Data Processing ──────────────────────────────────────────────────────────
pandas>=2.2.2
numpy>=1.26.4
pyarrow>=16.0.0
polars>=0.20.22

# ─── Database & Storage ───────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.29
duckdb>=0.10.2
boto3>=1.34.84
botocore>=1.34.84
s3fs>=2024.3.1

# ─── Configuration ────────────────────────────────────────────────────────────
hydra-core>=1.3.2
omegaconf>=2.3.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1

# ─── Utilities ────────────────────────────────────────────────────────────────
loguru>=0.7.2
tenacity>=8.3.0
tqdm>=4.66.4
rich>=13.7.1
click>=8.1.7
httpx>=0.27.0
